requirements.txt
PyGithub
fastapi
uvicorn[standard]
